agents/risk_analyzer_agent/pyproject.toml
[project]
name = "risk_analyzer_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.16",
    "google-adk>=1.8.0",
    "httpx>=0.28.1",
    "pyportfolioopt>=1.5.6",
    "python-a2a>=0.5.9",
    "python-dotenv>=1.1.1",
    "twelvedata>=1.2.25",
    "yfinance>=0.2.65",
]
